% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{check_dataset}
\alias{check_dataset}
\title{Check if dataset is compatible with mvpa's tools}
\usage{
check_dataset(data = NULL, verbose = FALSE)
}
\arguments{
\item{data}{Dataframe.}

\item{verbose}{Shall informative messages about dataset integrity be printed (default FALSE).}
}
\value{
List highlighting the problematic valuesthat  have been found and informs about dataset structure. Tells if dataset is ready for downstream analysis.
}
\description{
Test dataframe (objects x variables) for missing, zero, invariant negative and categorical string values (eg. male / female).
}
\examples{
data <- data.frame(sex = c("m", "w", "w", "m", "m"),
 age = c(24, 12, 14, 12, 0),
 marital_status = c("m", "s", "s", "d", "d"),
 status = c("sick", "healthy", "healthy", "sick", "healthy"),
 na_vals = c(1, 3, NA, NA, NA),
 na_vals2 = c(2, "B", 2 , "C", 3)
 )
 check_dataset(data)
}
