% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{correlation}
\alias{correlation}
\title{Calculation of a correlation matrix}
\usage{
correlation(
  data = NULL,
  response = NULL,
  correlate = "vars_w_response",
  selection = NULL
)
}
\arguments{
\item{data}{Dataframe.}

\item{response}{Variable all other variables shall be correlated with.}

\item{correlate}{Correlate variables with a single other variable "vars_w_response" (default). Correlate "variables" or "objects".}

\item{selection}{Select variables and objects for correlation.}
}
\value{
List with the (square) matrix, containing Pearson's r values and stored parameters.
\item{correlation_matrix}{Matrix with Pearson's r values - Square matrix if \code{correlate} equals "variables" or "objects"}
\item{correlate}{Identical to input argument \code{correlate}}
\item{response}{Identical to input argument \code{response}}
}
\description{
Correlate variables with each other (= correlation matrix) or select a response all variables shall correlated with.
}
\examples{
corr_list <- correlation(HOMA_IR, response = "HOMA_IR",
                         selection = c("Sex", "PA_2500_2999", "PA_100_249"))

corr_list <- correlation(HOMA_IR, correlate = "variables")

# corr_list <- correlation(HOMA_IR, correlate = "objects")
}
