% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{data_summary}
\alias{data_summary}
\title{Create descriptive statistics summary}
\usage{
data_summary(data = NULL)
}
\arguments{
\item{data}{Dataframe with only numeric entries.}
}
\value{
Dataframe with min, max, mean, median, variance and standard deviation as columns and all variables as rows.
\item{min}{Minimum}
\item{max}{Maximum}
\item{mean}{Mean}
\item{median}{Median}
\item{variance}{Variance}
\item{sd}{Standard deviation (N - 1)}
}
\description{
Data summary will created for dataframes that only consist of numerical values.
}
\examples{
data_summary(HOMA_IR)
}
