% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{enhance_dataset}
\alias{enhance_dataset}
\title{Join two dataframes based on rownames}
\usage{
enhance_dataset(data = NULL, data_to_add = NULL)
}
\arguments{
\item{data}{Dataframe.}

\item{data_to_add}{Dataframe with one or more columns that shall be added.}
}
\value{
Enhanced dataframe.
}
\description{
This function facilitates joining variables obtained downstream (principal component or PLS scores, etc.)
with the original dataframe. Ideal if original variables shall be substituted with latent variables for
the purpose of dimension reduction.
}
\examples{

# enhanced_dataset <- enhance_dataset(data = old_dataset, data_to_add = new_variables)
}
