% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute missing values}
\usage{
impute_missing_values(data = NULL, method = "mean", custom_value = NULL)
}
\arguments{
\item{data}{Dataframe with missing values (NAs).}

\item{method}{Imputation method: "mean" (default), "median", "custom". Mean and median values will be calculated by variable.}

\item{custom_value}{Value to substitute NAs in dataframe with.}
}
\value{
Dataframe with missing value substituted by defined method.
}
\description{
Missing values are calculated based on the selected method ('mean' or 'median') or are selected manually
by choosing a custom value.
}
\examples{

data <- data.frame(owns = c(-24, 125, 250, -150, 134),
                   na_vals = c(3, 5, 5, 3, NA),
                   na_vals2 = c(3, 3, 5 , NA, NA))
mean_imp_data <- impute_missing_values(data) # default method mean

# Dataframe-wide imputation with small value
custom_imp_data <- impute_missing_values(data, method = "custom", custom_value = 1e-6)
}
