% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{introduce_dummies}
\alias{introduce_dummies}
\title{Introduce dummy variables}
\usage{
introduce_dummies(data = NULL, col_names = NULL, add_indicator = FALSE)
}
\arguments{
\item{data}{Dataframe.}

\item{col_names}{(Optional) If only specific columns shall be converted to dummy columns.}

\item{add_indicator}{Add level name to variable name, separated by an underscore,  indicating 1's.
Only affects binary variables (default FALSE). Variables with more than two levels will get an indicator regardless.}
}
\value{
Dataframe with new dummy columns. Former character / factor-containing columns are removed.
}
\description{
Introduces binary dummy variables for columns that contain characters / factors.
In case a variable only contains two levels, only one column is retained to avoid redundancy.
The respective columns containing characters / factors are removed.
}
\examples{
# df <- introduce_dummies(df)
# df <- introduce_dummies(df, col_names = c("Gender", "marital_status"))
}
