% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{introduce_offset}
\alias{introduce_offset}
\title{Introduce offset value}
\usage{
introduce_offset(data = NULL, new_min = 1, col_names = NULL)
}
\arguments{
\item{data}{Dataframe.}

\item{new_min}{New (positive) minimum value (default = 1)}

\item{col_names}{Names of specific columns to be changed.}
}
\value{
Dataframe ready for scaling / transformation (eg. mvpa::scale_data).
}
\description{
Set offset to avoid negative and zero variable values. Useful when log or square root methods shall be applied.
}
\examples{
data <- data.frame(status = c("sick", "healthy", "healthy", "sick", "healthy"),
                   height = c(124, 142, 152, 163, 146),
                   owns = c(-24, 125, 250, -150, 134))

introduce_offset(data, new_min = 1)
}
