% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{introduce_unique_combinations}
\alias{introduce_unique_combinations}
\title{Introduce unique cross terms}
\usage{
introduce_unique_combinations(
  data = NULL,
  method = "*",
  ignore_variables = NULL
)
}
\arguments{
\item{data}{Dataframe with numeric entries.}

\item{method}{Operation to combine columns. Default is "*" to form products of column values.}

\item{ignore_variables}{Mention variables that all shall be ignored when forming combinations. Default is NULL.}
}
\value{
Enhanced dataframe that contains unique column combinations.
}
\description{
Based on the selected method (default is multiplication - "*"), unique variable combinations are formed and are added
to the dataframe.
}
\examples{
enhanced_data <- introduce_unique_combinations(HOMA_IR[1:3, 1:8])

}
