% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{keep_variables}
\alias{keep_variables}
\title{Keep selected variables}
\usage{
keep_variables(data = NULL, vars_to_keep = NULL)
}
\arguments{
\item{data}{Dataframe.}

\item{vars_to_keep}{Variable name(s) that shall be kept in the dataframe. The rest will be removed.}
}
\value{
Dataframe with kept variables only.
}
\description{
Selected variables are kept while the others are removed.
}
\examples{
# df <- keep_variables(df, vars_to_keep = c("Age", "Skinfold"))
# df <- keep_variables(df, "Gender")
}
