% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_projection.R
\name{perform_covariate_projection}
\alias{perform_covariate_projection}
\title{Remove covariate influence from multivariate data}
\usage{
perform_covariate_projection(
  X_aug = NULL,
  covariates = NULL,
  standardize = FALSE,
  reorder_by_covariates = TRUE
)
}
\arguments{
\item{X_aug}{Dataframe that contains explanatory (X), response (y) and potential covariate (Z) variables.}

\item{covariates}{Names of covariate columns.}

\item{standardize}{Boolean to indicate standardization X_standardized <- X / X_sd (default FALSE).}

\item{reorder_by_covariates}{Boolean to indicate if order of selected covariates shall be reflected in result dataframe order (default TRUE).}
}
\value{
A List containing (1) a residual variance dataframe and (2) a dataframe with explained variance by the covariates.
}
\description{
More detailed description
}
\examples{
# res <- perform_covariate_projection(X_aug, covariates = c("Gender", "Age"), standardize = TRUE)
}
