% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_mc_sr.R
\name{perform_mc_pls_tp}
\alias{perform_mc_pls_tp}
\title{PLS-R model building and Target Projection}
\usage{
perform_mc_pls_tp(
  data = NULL,
  response = NULL,
  nr_components = 8,
  mc_resampling = TRUE,
  cal_ratio = 0.5,
  nr_repetitions = 100,
  standardize = FALSE,
  use_sd_full_dataset = FALSE,
  validation_threshold = 0.5,
  cost_function_type = "rmsep"
)
}
\arguments{
\item{data}{Dataframe that contains explanatory variables (X) and a response variable (y).}

\item{response}{Name of the response variable.}

\item{nr_components}{Number of components to test (default 8).}

\item{mc_resampling}{Boolean to indicate usage of Monte Carlo resampling (default TRUE). If FALSE, nr_of components will be used directly to build the PLS-R model.}

\item{cal_ratio}{Ratio of calibration data to total available data (eg. 0.8 -> 80\% of the data is used for calibration).}

\item{nr_repetitions}{Number of Monte Carlo repetitions (default 100).}

\item{standardize}{Boolean to indicate standardization X_standardized <- X / X_sd (default FALSE).}

\item{use_sd_full_dataset}{Boolean to indicate if the standard deviation (n -1) of the full dataset shall be used. Default FALSE. Recommended/Necessary when Monte Carlo resampling leads to samples with variables that exhibit zero variance.}

\item{validation_threshold}{Threshold fraction upon a component number is selected. Based on Monte Carlo resampling. Too high values lead to overfitting (default 0.5).}

\item{cost_function_type}{Cost function to use to measure model performance. Root mean squared error ("rmsep", default) and mean absolute error ("mae") available.}
}
\value{
A list that contains the following results:
\item{errors}{\code{cost_function} error value matrix with dimensions \code{nr_repetitions} * \code{nr_components}}
\item{cost_function}{Indicates with cost function was used (RMSEP, MAE)}
\item{A_optimal}{Optimal number of components. Component count was used to perform target projection on full dataset}
\item{fractions}{Ratio of cost function values (higher component) above and below the median cost function value of the lesser component. The selected \code{validation_threshold} determines \code{A_optimal}}
\item{tested_components}{List with target projection values of tested components 1 to \code{nr_components} for each repetition}
\item{target_projection_A_opt}{List with target projection values of the full dataset and based on \code{A_optimal} retained components}
\item{explained_variance_in_X}{Explained variance in X for each component - Multiply by 100 to obtain \%}
\item{explained_variance_in_y}{Explained variance in y for each component - Multiply by 100 to obtain \%}
\item{explained_variance_tTP}{Explained variance of target score tTP - Multiply by 100 to obtain \%}
}
\description{
This function allows PLS model building, using a combined dataframe with explanatory and a single response variable.
It is recommended to use Monte Carlo resampling to determine an optimal number of PLS components used for the model.
During Monte Carlo resampling, a fixed number of random objects is picked to build PLS models with varying numbers of
PLS components. A loss function (RMSEP by default) calculates the goodness of the built models and the objects left out
for validation. The model with the most robust and best performance (= lowest cost function value) indicates the number of components
to use for the full dataset.
}
\examples{
# perform_mc_pls_tp(data = HOMA_IR, response = "HOMA_IR", nr_components = 8,
#                   mc_resampling = TRUE, cal_ratio = 0.5, repetitions = 100, standardize = TRUE)
}
