% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{perform_pca}
\alias{perform_pca}
\title{Perform principal component analysis}
\usage{
perform_pca(data = NULL, standardize = FALSE, exclude_variables = NULL)
}
\arguments{
\item{data}{Dataframe.}

\item{standardize}{Shall data be standardized (default False).}

\item{exclude_variables}{Variables to exclude from principal component analysis.}
}
\value{
List that contains the following values:
\item{scores}{Matrix with PCA scores}
\item{loadings}{Matrix with PCA loadings}
\item{variances}{Variance of each principal component}
\item{explained_variance}{Explained variance by each component}
}
\description{
The function performs principal component analysis based on singular value decomposition (SVD).
As a consequence, no missing values are allowed. Use the \code{check_dataset(dataset))} function
to check if the dataset is in suited.
}
\examples{
# result <- perform_pca(HOMA_IR)
}
