% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_mc_sr.R
\name{plot_cost_function}
\alias{plot_cost_function}
\title{Plot cost function values}
\usage{
plot_cost_function(
  result_list = NULL,
  validation_threshold = 0.5,
  show_confidence_limits = FALSE,
  confidence_limits = c(0.05, 0.95),
  rel_font_size = 1
)
}
\arguments{
\item{result_list}{Result list obtained from \code{perform_mc_pls_tp()}.}

\item{validation_threshold}{Threshold fraction upon a component number is selected. Based on Monte Carlo resampling. Too high values lead to overfitting (default 0.5).}

\item{show_confidence_limits}{Show confidence limits (default FALSE).}

\item{confidence_limits}{Confidence limits. Default 5\% and 95\% (-> c(.05, .95)).}

\item{rel_font_size}{Relative font size based on default font size 11 (default 1).}
}
\value{
Bar plot that shows cost function values (default RMSEP) for each tested component. Error bars indicate 5% and 95% confidence limit.
}
\description{
This function uses the results from \code{perform_mc_pls_tp()} and plots the cost function values (eg. RMSEP) on the y-axis
and the tested PLS components on the x-axis.
}
\examples{
# # result_list obtained from perform_mc_pls_tp()
# cost_function_plot(result_list, validation_threshold = 0.5)
}
