% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_projection.R
\name{plot_explained_var_by_covariates}
\alias{plot_explained_var_by_covariates}
\title{Plot explained variance by covariate}
\usage{
plot_explained_var_by_covariates(
  covariate_analysis_result = NULL,
  x_filter = NULL,
  rel_font_size = 1,
  rotate = FALSE
)
}
\arguments{
\item{covariate_analysis_result}{Result list, obtained from perform_covariate_projection().}

\item{x_filter}{Filter for variables / features.}

\item{rel_font_size}{Relative font size based on default font size 11 (default 1).}

\item{rotate}{Rotate plot by 90 degrees to read axis labels more easily.}
}
\value{
Stacked bar plot showing the residual variance and explained variance by covariates.
}
\description{
Plot explained variance by covariate
}
\examples{
# # Perform covariate projection
# res <- perform_covariate_projection(X_aug, covariates = c("Gender", "Age"), standardize = TRUE)
# plot_explained_var_by_covariates(res$explained_variance_by_covariates)
}
