% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca_2d}
\alias{plot_pca_2d}
\title{Plot 2D principal component analysis results}
\usage{
plot_pca_2d(
  pca_result,
  plot = "scores",
  PCs_to_plot = c(1, 2),
  rel_point_size = 1,
  rel_font_size = 1
)
}
\arguments{
\item{pca_result}{PCA result obtained from perform_pca() method.}

\item{plot}{Either "loadings" or "scores" (default).}

\item{PCs_to_plot}{Tuple of PCs to plot (default c(1, 2)).}

\item{rel_point_size}{Relative point size (default 1, point size 3).}

\item{rel_font_size}{Relative font size (default 1, font size 11).}
}
\value{
ggplot2 2D plot with PCA scores or loadings.
}
\description{
Choose between PCA "scores" and "loadings" and plot the PCA results.
}
\examples{

pca <- perform_pca(HOMA_IR)

plot_pca_2d(pca, "scores", c(2, 3))
}
