% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca_value}
\alias{plot_pca_value}
\title{Plot absolute scores or loadings of a specific principal component in form of a bar plot.}
\usage{
plot_pca_value(
  pca_result = NULL,
  plot = "scores",
  PC_to_plot = 1,
  y_filter = NULL,
  x_filter = NULL,
  rel_font_size = 1,
  rotate = FALSE
)
}
\arguments{
\item{pca_result}{PCA result obtained from perform_pca() method.}

\item{plot}{Either "loadings" or "scores" (default).}

\item{PC_to_plot}{Principal component (PC) to plot (default 1).}

\item{y_filter}{Minimum and Maximum filter for values. Either single value or vector c(lower, higher).}

\item{x_filter}{Filter for objects / variables.}

\item{rel_font_size}{Relative font size (default 1, font size 11).}

\item{rotate}{Rotate plot by 90 degrees to read axis labels more easily.}
}
\value{
Bar plot with desired Principal Component Analysis scores or loading values.
}
\description{
Plot absolute scores or loadings of a specific principal component in form of a bar plot.
}
\examples{

pca <- perform_pca(HOMA_IR)

plot_pca_value(pca, "scores", 2)
}
