% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_scree_plot}
\alias{plot_scree_plot}
\title{Plot scree plot}
\usage{
plot_scree_plot(
  pca_result,
  how_many_PCs = 10,
  add_cumulative_perc = FALSE,
  rel_font_size = 1
)
}
\arguments{
\item{pca_result}{PCA result obtained from perform_pca() method.}

\item{how_many_PCs}{How many PCs should be shown in the plot.}

\item{add_cumulative_perc}{Add optional cumulative percentage line (default FALSE).}

\item{rel_font_size}{Relative font size (default 1, font size 11).}
}
\value{
ggplot scree plot.
}
\description{
The scree plot indicates the explained variance per principal component.
An optional line can be added to the plot that indicates the cumulative percentage plot,
indicating the increase in explained variance for each principal component.
}
\examples{
pca <- perform_pca(HOMA_IR)

plot_scree_plot(pca, 4, TRUE, TRUE)
}
