% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_mc_sr.R
\name{plot_tp_value}
\alias{plot_tp_value}
\title{Plot target projection values (full dataset)}
\usage{
plot_tp_value(
  result_list = NULL,
  tp_value_to_plot = "selectivity_ratio",
  y_filter = NULL,
  x_filter = NULL,
  rel_font_size = 1,
  rotate = FALSE
)
}
\arguments{
\item{result_list}{Result list obtained from \code{perform_mc_pls_tp()}.}

\item{tp_value_to_plot}{Target projection value: "wTP", "tTP", "pTP", "selectivity_ratio", "selectivity_fraction", "MCorrC" (multivariate correlation coefficient) and "MCovC" (multivariate covariance coefficient). See examples for details.}

\item{y_filter}{Minimum and Maximum filter for values. Either single value or vector c(lower, higher).}

\item{x_filter}{Filter for variables / features.}

\item{rel_font_size}{Relative font size based on default font size 11 (default 1).}

\item{rotate}{Rotate plot by 90 degrees to read axis labels more easily.}
}
\value{
Bar plot with desired Target Projection variable.
}
\description{
Plot Target Projection variables, such as scores, loadings, weights, selectivity ratio, selectivity fraction,
multivariate correlation coefficients (MCorrC) and unstandardized multivariate correlation coefficients (MCovC)
}
\examples{
# # Obtain optimal PLS-R model and perform Target Projection on full dataset,
# # using the optimal number of components.
# result_list <- perform_mc_pls_tp(data = HOMA_IR, response = "HOMA_IR", nr_components = 8,
#                                 cal_ratio = 0.5, repetitions = 100, standardize = TRUE)

# plot_tp_value(result_list)
}
