% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_mc_sr.R
\name{plot_tp_value_mc}
\alias{plot_tp_value_mc}
\title{Plot sampled target projection values}
\usage{
plot_tp_value_mc(
  result_list = NULL,
  tp_value_to_plot = "selectivity_ratio",
  component = NULL,
  confidence_limits = c(0.025, 0.975),
  x_filter = NULL,
  y_filter = NULL,
  rel_font_size = 1,
  rotate = FALSE
)
}
\arguments{
\item{result_list}{Result list obtained from \code{perform_mc_pls_tp()}.}

\item{tp_value_to_plot}{Target projection value: "wTP", "tTP", "pTP", "selectivity_ratio" to "sr3", MCorrC (multivariate correlation coefficient) and unstandardized MCorrC . See examples for details.}

\item{component}{Component number used for PLS regression and subsequent Target Projection.}

\item{confidence_limits}{Confidence limits. Default 2.5\% and 97.5\% (-> c(.025, .975)).}

\item{x_filter}{Filter for variables / features.}

\item{y_filter}{Minimum and Maximum filter for values. Either single value or vector c(lower, higher).}

\item{rel_font_size}{Relative font size based on default font size 11 (default 1).}

\item{rotate}{Rotate plot by 90 degrees to read axis labels more easily.}
}
\value{
Bar plot with Target projection value for each feature. Error bars indicate confidence limits.
}
\description{
Plot Target Projection variables, such as scores, loadings, weights, selectivity ratio, selectivity fraction,
multivariate correlation coefficients (MCorrC) and unstandardized multivariate correlation coefficients (MCovC).
In addition, confidence limits are provided which are derived from the Monte Carlo resampling.
}
\examples{
# # result_list obtained from perform_mc_pls_tp()
# plot_tp_value_mc(result_list, tp_value_to_plot = "selectivity_fraction", component = 3)
}
