% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_mc_sr.R
\name{plot_variable_variance_distribution}
\alias{plot_variable_variance_distribution}
\title{Plot variable variance distribution based}
\usage{
plot_variable_variance_distribution(
  result_list = NULL,
  x_filter = NULL,
  rel_font_size = 1,
  rotate = FALSE
)
}
\arguments{
\item{result_list}{Result list obtained from perform_mc_pls_tp().}

\item{x_filter}{Filter for variables / features.}

\item{rel_font_size}{Relative font size based on default font size 11 (default 1).}

\item{rotate}{Rotate plot by 90 degrees to read axis labels more easily.}
}
\value{
Stacked bar plot with explained and residual variance per feature based on target projection.
}
\description{
This stacked bar plot depicts explained and residual variance for each variable.
Explained and residual variance are obtained through target projection.
}
\examples{
# # Obtain optimal PLS-R model and perform Target Projection on full dataset,
# # using the optimal number of components.
# result_list <- perform_mc_pls_tp(data = HOMA_IR, response = "HOMA_IR", nr_components = 8,
#                                 cal_ratio = 0.5, repetitions = 100, standardize = TRUE)

# plot_variable_variance_distribution(result_list)
}
