% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ-plot to check for normality}
\usage{
qq_plot(data = NULL, vars_to_plot = NULL, nr_cols = 5, rel_font_size = 1)
}
\arguments{
\item{data}{Dataframe.}

\item{vars_to_plot}{Selection of variables. If NULL (default) the the first variable will be shown.}

\item{nr_cols}{Number of QQ-plots per line.}

\item{rel_font_size}{Relative font size based on default font size 11 (default 1).}
}
\value{
QQ-plot(s) of variables.
}
\description{
Builds quantile-quantile plots based on selected variables. Several variables can be selected simultaneously.
}
\examples{
qq_plot(HOMA_IR, c("PA_100_249", "Age"))
}
