% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{remove_objects}
\alias{remove_objects}
\title{Remove objects / samples from dataframe.
Requires column with character-based identifier.}
\usage{
remove_objects(data = NULL, objs_to_remove = NULL, remove_na = FALSE)
}
\arguments{
\item{data}{Dataframe}

\item{objs_to_remove}{Objects that shall be removed from the dataframe. Is ignored, if remove_na is TRUE.}

\item{remove_na}{Remove all objects that contain NAs (default FALSE).}
}
\value{
Dataframe with removed objects.
}
\description{
Remove objects / samples from dataframe.
Requires column with character-based identifier.
}
\examples{
# # Add unique id to dataframe that is based on row indices
# df <- add_unique_id(df)
# df <- remove_objects(df, c("0011", "0231", "2357"))
# df <- remove_objects(df, "0042")
}
