% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{remove_variables}
\alias{remove_variables}
\title{Remove variables / features from dataframe}
\usage{
remove_variables(
  data = NULL,
  vars_to_remove = NULL,
  remove_na = FALSE,
  remove_invariants = FALSE
)
}
\arguments{
\item{data}{Dataframe.}

\item{vars_to_remove}{Variable name(s) that shall be removed from the dataframe.}

\item{remove_na}{Remove all objects that contain NAs (default FALSE).}

\item{remove_invariants}{Remove variables that are invariant (variance = 0).}
}
\value{
Dataframe with removed variables.
}
\description{
Remove variables / features from dataframe
}
\examples{
# df <- remove_variables(df, vars_to_remove = c("Age", "Skinfold"))
# df <- remove_variables(df, "Gender")
}
