% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{scale_data}
\alias{scale_data}
\title{Scale / Normalize dataset}
\usage{
scale_data(data = NULL, col_names = NULL, method = "log10", n_for_root = 3)
}
\arguments{
\item{data}{Dataframe.}

\item{col_names}{(Optional) Names with numeric columns that shall be scaled / normalized}

\item{method}{Methods to apply to data: "log10", "log2", "ln", "sqrt", "nth_root", "normalize_to_max", "min_max", "sd" (standardization), (default "log10").}

\item{n_for_root}{Integer to select nth-root (eg. 3 -> cubic root).}
}
\value{
Scaled dataframe.
}
\description{
Requires that dataset is cleaned from negative values, in case log or nth-root scaling is selected.
}
\examples{
data <- data.frame(status = c("sick", "healthy", "healthy", "sick", "healthy"),
                   height = c(124, 142, 152, 163, 146),
                   owns = c(-24, 125, 250, -150, 134))

data <- introduce_offset(data)

scale_data(data, method = "nth_root", n_for_root = 4)

scale_data(data, col_names = "height", method = "min_max")

scale_data(data, col_names = c("height", "owns"), "normalize_to_max")
}
