% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_mc_sr.R
\name{target_projection}
\alias{target_projection}
\title{Target projection on latent variable regression model}
\usage{
target_projection(
  X = NULL,
  y = NULL,
  X_sd = NULL,
  y_sd = NULL,
  b_LVR = NULL,
  sr_only = FALSE
)
}
\arguments{
\item{X}{Matrix object with explanatory variables (n x m).}

\item{y}{Matrix object with response variables (n x 1).}

\item{X_sd}{Standard deviation of X - Used for calculation of MCovC}

\item{y_sd}{Standard deviation of y - Used for calculation of MCovC}

\item{b_LVR}{Matrix object with regression coefficients (m x 1).}

\item{sr_only}{Boolean to indicate if only selectivity ratios and derived values shall be returned (default FALSE).}
}
\value{
List that contains the following values:
\item{wTP}{Target projection weights}
\item{tTP}{Target projection scores}
\item{pTP}{Target projection loadings}
\item{Etp}{Target projection residual matrix}
\item{explained_variance}{Explained variance - derived from \code{tTP} and \code{pTP}}
\item{residual_variance}{Residual variance - derived from \code{Etp}}
\item{explained_variance_tTP}{Variance explained by the target score tTP = \eqn{\frac{explained\:variance}{(exaplained\:variance\:+\:residual\:variance)}}}
\item{selectivity_ratio}{Selectivity ratio (SR) = \eqn{\frac{explained\:variance}{residual\:variance}}}
\item{selectivity_fraction}{Selectivity fraction (SF) = \eqn{\frac{explained\:variance}{var(X)}}}
\item{MCorrC}{Multivariate correlation coefficient - MCorrC = \eqn{\frac{pTP}{|pTP|} * \sqrt{|SF\:R2|}}}
\item{MCovC}{Multivariate covariance coefficient - MCovC = \eqn{\frac{MCorrC}{sd(X)}}}
}
\description{
This function calculates all target projection values, including all selectivity ratios and derivatives.
A detailed algebraic description of the target projection can be found in
Kvalheim, e3211, 2020, Journal of Chemometrics (\href{doi.org/10.1002/cem.3211}{link to publication}).
}
\examples{
# target_projection(X, y, b_LVR, TRUE, X_sd, y_sd)
}
